+++ tests/integration/components/flexberry-validationsummary-test.js
@@ -1,6 +1,7 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 import Errors from 'ember-validations/errors';
+import Ember from 'ember';
 
 moduleForComponent('flexberry-validationsummary', 'Integration | Component | flexberry validationsummary', {
   integration: true
@@ -55,4 +56,54 @@ test('it should be visible if errors presence', function (assert) {
 
   this.render(hbs`{{flexberry-validationsummary errors=errors}}`);
   assert.equal(this.$(':first-child').is(':visible'), true);
+});
+
+test('it renders changing error list', function (assert) {
+  let errors = Errors.create();
+  Ember.set(errors, 'testProperty', ['error1']);
+  Ember.set(this, 'errors', errors);
+
+  // Initialize by errors property.
+  this.render(hbs`{{flexberry-validationsummary errors=errors}}`);
+  assert.equal(this.$().text().trim(), 'error1');
+
+  let testPropertyErrorList = new Ember.A(['error11', 'error12']);
+  Ember.run(() => {
+    // Change existing errors property.
+    Ember.set(errors, 'testProperty', testPropertyErrorList);
+  });
+
+  assert.equal(this.$().text().replace(/\n|\s/g, ""), 'error11error12');
+
+  Ember.run(() => {
+    // Add new errors property.
+    Ember.set(errors, 'testProperty2', ['error2']);
+  });
+  
+  assert.equal(this.$().text().replace(/\n|\s/g, ""), 'error11error12error2');
+
+  Ember.run(() => {
+    // Remove element from existing errors property.
+    Ember.get(errors, 'testProperty').popObject();
+  });
+
+  assert.equal(this.$().text().replace(/\n|\s/g, ""), 'error11error2');
+
+  Ember.run(() => {
+    // Add element to existing errors property.
+    Ember.get(errors, 'testProperty').pushObject('error13');
+  });
+
+  assert.equal(this.$().text().replace(/\n|\s/g, ""), 'error11error13error2');
+
+  let errors2 = Errors.create();
+  Ember.set(errors2, 'testProperty2', ['error22']);
+  Ember.set(errors2, 'testProperty3', ['error3']);
+
+  Ember.run(() => {
+    // Change errors object.
+    Ember.set(this, 'errors', errors2);
+  });
+
+  assert.equal(this.$().text().replace(/\n|\s/g, ""), 'error22error3');
 });